extract unary to file
